Create a singleton client for all tests
