+++ yaml_test_runner/Cargo.toml
@@ -17,6 +17,7 @@ itertools = "0.8.2"
 Inflector = "0.11.4"
 lazy_static = "1.4.0"
 log = "0.4.8"
+once_cell = "1.4.0"
 path-slash = "0.1.1"
 quote = "~0.3"
 regex = "1.3.1"
