+++ yaml_test_runner/tests/common/mod.rs
@@ -20,4 +20,3 @@
 pub mod macros;
 pub mod client;
 pub mod transform;
-pub mod util;
